Moves RootsTracker to submodule
